move bubble_sort and reverse_order to sorting/mixins.py. Improve documentation and code coverage.
